TCAP: resolve DHA / CCO dependencies on TC-U initiated BEGIN
Prior to this patch, DHA needs to know the CCO pid, and CCO needed to
know the DHA pid at startup.  The idea was to start CCO from within the
init() callback in DHA.  However, this caused a deadlock in the
supervisor.

We now use the locally registred process names (tcap_cco_DLGID) to
resolve the process.  However, it might be cleaner to use an ets table
at some later point.  For now it is nice to see the name of each process
in pman for debugging...